# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Renta de Apartamentos - API",
        "description": "Documentación de la API del sistema de rentas de apartamentos. Permite gestionar usuarios, apartamentos, quejas, contratos y más.",
        "contact": {
            "email": "tucorreo@ejemplo.com"
        },
        "license": {
            "name": "Todos los derechos reservados - Universidad Piloto de Colombia"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/permisos": {
            "get": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Listar todos los permisos",
                "operationId": "8feceb78c9ff332ad552bb98ae0657d3",
                "responses": {
                    "200": {
                        "description": "Lista de permisos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Permiso"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Crear un nuevo permiso",
                "operationId": "f5312187af2d66fda3392fd3222089fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Permiso"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Permiso creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Permiso"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/permisos/{id}": {
            "get": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Obtener un permiso por ID",
                "operationId": "36d51e43024849e1a739a7021da163ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permiso encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Permiso"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Actualizar un permiso",
                "operationId": "bc6245339799a3f8a9a81e93faa293b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Permiso"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permiso actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Permiso"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Eliminar un permiso",
                "operationId": "1b8b8d685cca25f33209f68652f78fb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permiso eliminado"
                    }
                }
            }
        },
        "/api/permisos/{id}/edit": {
            "get": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Mostrar datos de un permiso para editar",
                "operationId": "1adc877089ced3e56060d125f8cfea72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del permiso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Permiso"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Permiso": {
                "title": "Permiso",
                "description": "Modelo de permiso del sistema",
                "properties": {
                    "id": {
                        "description": "ID único del permiso",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre del permiso",
                        "type": "string",
                        "example": "Crear usuarios"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}